refactor vector tap
