+++ lib/vector-lib/src/lib.rs
@@ -9,7 +9,7 @@ pub use vector_buffers as buffers;
 pub use vector_common::event_test_util;
 pub use vector_common::{
     assert_event_data_eq, btreemap, byte_size_of, byte_size_of::ByteSizeOf, conversion,
-    encode_logfmt, finalization, finalizer, impl_event_data_eq, internal_event, json_size,
+    encode_logfmt, finalization, finalizer, id, impl_event_data_eq, internal_event, json_size,
     registered_event, request_metadata, sensitive_string, shutdown, trigger, Error, Result,
     TimeZone,
 };
