+++ lib/vector-tap/Cargo.toml
@@ -7,7 +7,10 @@ publish = false
 license = "MPL-2.0"
 
 [dependencies]
+async-graphql = { version = "7.0.7", default-features = false, features = ["playground"], optional = true}
 colored = { version = "2.1.0", default-features = false }
+futures = { version = "0.3.30", default-features = false, features = ["std"] }
+glob = { version = "0.3.1", default-features = false }
 serde_yaml = { version = "0.9.34", default-features = false }
 snafu = { version = "0.7.5", default-features = false }
 tokio = { version = "1.39.3", default-features = false, features = ["time"] }
@@ -15,7 +18,11 @@ tokio-stream = { version = "0.1.15", default-features = false, features = ["sync
 tokio-tungstenite = { version = "0.20.1", default-features = false }
 tracing = { version = "0.1.34", default-features = false }
 url = { version = "2.5.2", default-features = false }
+uuid.workspace = true
 vector-api-client = { path = "../vector-api-client" }
+vector-common = { path = "../vector-common" }
+vector-core = { path = "../vector-core" }
+vector-buffers = { path = "../vector-buffers" }
 futures-util = "0.3.30"
 
 [dev-dependencies]
@@ -23,6 +30,9 @@ chrono = { workspace = true }
 portpicker = { path = "../portpicker" }
 serde_json = { workspace = true }
 tokio = { version = "1.39.3", default-features = false, features = ["test-util"] }
+
+[features]
+api = ["dep:async-graphql"]
