+++ src/api/schema/events/encoding.rs
@@ -2,7 +2,7 @@ use async_graphql::Enum;
 
 #[derive(Enum, Copy, Clone, PartialEq, Eq)]
 /// Encoding format for the event
-pub(crate) enum EventEncodingType {
+pub enum EventEncodingType {
     Json,
     Yaml,
     Logfmt,
