+++ src/api/schema/events/log.rs
@@ -3,10 +3,11 @@ use std::borrow::Cow;
 use async_graphql::Object;
 use chrono::{DateTime, Utc};
 use vector_lib::encode_logfmt;
+use vector_lib::event;
+use vector_lib::tap::topology::TapOutput;
 use vrl::event_path;
 
 use super::EventEncodingType;
-use crate::{event, topology::TapOutput};
 
 #[derive(Debug, Clone)]
 pub struct Log {
