+++ src/api/schema/events/trace.rs
@@ -1,9 +1,10 @@
 use async_graphql::Object;
 use vector_lib::encode_logfmt;
+use vector_lib::event;
+use vector_lib::tap::topology::TapOutput;
 use vrl::event_path;
 
 use super::EventEncodingType;
-use crate::{event, topology::TapOutput};
 
 #[derive(Debug, Clone)]
 pub struct Trace {
