+++ src/topology/mod.rs
@@ -20,13 +20,12 @@ mod task;
 mod test;
 
 use std::{
-    collections::{HashMap, HashSet},
     panic::AssertUnwindSafe,
     sync::{Arc, Mutex},
 };
 
 use futures::{Future, FutureExt};
-use tokio::sync::{mpsc, watch};
+use tokio::sync::mpsc;
 use vector_lib::buffers::topology::channel::{BufferReceiverStream, BufferSender};
 
 pub use self::builder::TopologyPieces;
@@ -35,7 +34,7 @@ pub use self::running::{RunningTopology, ShutdownErrorReceiver};
 
 use self::task::{Task, TaskError, TaskResult};
 use crate::{
-    config::{ComponentKey, Config, ConfigDiff, Inputs, OutputId},
+    config::{ComponentKey, Config, ConfigDiff},
     event::EventArray,
     signal::ShutdownError,
 };
@@ -46,34 +45,6 @@ type BuiltBuffer = (
     BufferSender<EventArray>,
     Arc<Mutex<Option<BufferReceiverStream<EventArray>>>>,
 );
-
-/// A tappable output consisting of an output ID and associated metadata
-#[derive(Debug, Clone, Hash, PartialEq, Eq)]
-pub struct TapOutput {
-    pub output_id: OutputId,
-    pub component_kind: &'static str,
-    pub component_type: String,
-}
-
-/// Resources used by the `tap` API to monitor component inputs and outputs,
-/// updated alongside the topology
-#[derive(Debug, Default, Clone)]
-pub struct TapResource {
-    // Outputs and their corresponding Fanout control
-    pub outputs: HashMap<TapOutput, fanout::ControlChannel>,
-    // Components (transforms, sinks) and their corresponding inputs
-    pub inputs: HashMap<ComponentKey, Inputs<OutputId>>,
-    // Source component keys used to warn against invalid pattern matches
-    pub source_keys: Vec<String>,
-    // Sink component keys used to warn against invalid pattern matches
-    pub sink_keys: Vec<String>,
-    // Components removed on a reload (used to drop TapSinks)
-    pub removals: HashSet<ComponentKey>,
-}
-
-// Watcher types for topology changes.
-type WatchTx = watch::Sender<TapResource>;
-pub type WatchRx = watch::Receiver<TapResource>;
 
 pub(super) fn take_healthchecks(
     diff: &ConfigDiff,
